# Replace `ListPostings.Seek`'s binary search call by the generic `slices.BinarySearch`