This is a hot path, and I saw it taking a while on some profiles.

First, I inlined the call, then on @bboreham suggestion I tried the generic version of the search, which turned out to be even faster (for some reason).

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb/index
                           │    main     │               inlined               │               generic               │
                           │   sec/op    │   sec/op     vs base                │   sec/op     vs base                │
Intersect/LongPostings1-12   30.38µ ± 1%   14.25µ ± 3%  -53.10% (p=0.000 n=10)   11.70µ ± 1%  -61.49% (p=0.000 n=10)
Intersect/LongPostings2-12   43.31m ± 2%   42.73m ± 2%        ~ (p=0.063 n=10)   43.35m ± 3%        ~ (p=0.315 n=10)
Intersect/ManyPostings-12    131.9m ± 2%   126.6m ± 2%   -4.04% (p=0.000 n=10)   133.7m ± 1%   +1.41% (p=0.004 n=10)
geomean                      5.578m        4.255m       -23.71%                  4.078m       -26.88%
```