@colega 2024-07-02
@bboreham also pointed out that it might help checking the immedately next posting before doing the complete binary search, as many intersect operations are realized on postings that are a subset of each other. I would like to do that in a follow up PR.